+++ CHANGELOG.md
@@ -9,9 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - Enzyme diffusion simulation.
+- Settings export to `gpseq_fromfish`.
 
 ### Changed
 - Moved distance functions to separate `tools` sub-module.
+- Now using `ggc` to confirm settings in `gpseq_anim`.
 
 ### Fixed
 - `return` bug crashing the pipeline when a condition has no nuclei.